splits solver into its own file modifies solution interfaces
