+++ 11_Robots/ghost_swerve_mpc_planner/CMakeLists.txt
@@ -32,27 +32,52 @@ endforeach()
 
 ament_export_dependencies(${DEPENDENCIES})
 
-# # Set Include Directories
-# set(INCLUDE
-#   include
-# )
+# Set Include Directories
+set(INCLUDE
+  include
+)
 
-# include_directories(${INCLUDE})
-# ament_export_include_directories(${INCLUDE})
+include_directories(${INCLUDE})
+ament_export_include_directories(${INCLUDE})
 
+###################
+#### Libraries ####
+###################
+add_library(swerve_mpc_planner SHARED
+  src/swerve_mpc_planner.cpp
+)
+target_include_directories(swerve_mpc_planner
+  PUBLIC
+  $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+  $<INSTALL_INTERFACE:include>)
+ament_target_dependencies(swerve_mpc_planner
+  ${DEPENDENCIES}
+)
+target_link_libraries(swerve_mpc_planner
+  casadi
+  matplotlib_cpp::matplotlib_cpp
+  ${PYTHON_LIBRARIES}
+)
+ament_export_targets(swerve_mpc_planner HAS_LIBRARY_TARGET)
+install(
+  TARGETS swerve_mpc_planner
+  EXPORT swerve_mpc_planner
+  LIBRARY DESTINATION lib
+  ARCHIVE DESTINATION lib
+  RUNTIME DESTINATION bin
+  INCLUDES DESTINATION include
+)
 
-# ##################
-# ## Executables ###
-# ##################
+###################
+### Executables ###
+###################
 # Main to generate Problem Formulation for MDS MPC
 add_executable(swerve_mpc_planner_node src/swerve_mpc_planner_node.cpp)
 ament_target_dependencies(swerve_mpc_planner_node
   ${DEPENDENCIES}
 )
 target_link_libraries(swerve_mpc_planner_node
-  casadi
-  matplotlib_cpp::matplotlib_cpp
-  ${PYTHON_LIBRARIES}
+  swerve_mpc_planner
 )
 target_include_directories(swerve_mpc_planner_node
   PUBLIC
