+++ 11_Robots/ghost_swerve_mpc_planner/config/casadi_swerve_config.yaml
@@ -1,7 +1,7 @@
 swerve_mpc_planner_node:
   ros__parameters:
-    time_horizon: 2.0
-    dt: 0.01
+    time_horizon: 5.0
+    dt: 0.05
     num_swerve_modules: 4
 
     chassis_width: 16.5
@@ -13,12 +13,12 @@ swerve_mpc_planner_node:
     wheel_constraint_tolerance: 0.001
 
     default_state_weights:
-      - 0.0 # base_pose_x
-      - 0.0 # base_pose_y
-      - 0.0 # base_pose_theta
-      - 1000.0 # base_vel_x
-      - 1000.0 # base_vel_y
-      - 1000.0 # base_vel_theta
+      - 10000.0 # base_pose_x
+      - 10000.0 # base_pose_y
+      - 10000.0 # base_pose_theta
+      - 0.0 # base_vel_x
+      - 0.0 # base_vel_y
+      - 0.0 # base_vel_theta
       - 0.0 # base_accel_x
       - 0.0 # base_accel_y
       - 0.0 # base_accel_theta
